chore: 레포지토리 구현체와 인터페이스 분리
