+++ Popcorn-iOS/Source/Domain/Interfaces/Repositories/Login/LoginRepositoryProtocol.swift
@@ -1,0 +1,10 @@
+//
+//  LoginRepositoryProtocol.swift
+//  Popcorn-iOS
+//
+//  Created by 김성훈 on 1/28/25.
+//
+
+protocol LoginRepositoryProtocol {
+    func login(username: String, password: String, completion: @escaping (Result<Token, Error>) -> Void)
+}
